✨ feat: Added request sending and shared structure separation
