+++ crates/cli/Cargo.toml
@@ -5,4 +5,7 @@ edition = "2021"
 
 [dependencies]
 clap = { version = "4.5.32", features = ["derive"] }
+futures = "0.3.31"
 regex = "1.11.1"
+shared = { path = "../shared" }
+client = { path = "../client" }
