+++ crates/cli/src/lib.rs
@@ -41,5 +41,3 @@ pub async fn init() {
         Commands::Uninstall {} => {}
     }
 }
-
-pub use parsers::Package;
