+++ crates/shared/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "shared"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+regex = "1.11.1"
